Add files via upload
This project aims to explore the impact of COVID-19 deaths worldwide. By analyzing various factors and patterns surrounding the pandemic's consequences, we seek to gain insights into the global magnitude of the crisis. Through data exploration, we aim to contribute to a better understanding of the broader repercussions on societies globally.