Add files via upload
The project focuses on conducting data cleaning of Real Estate data using SQL. Data cleaning is an essential step in the data analysis process, as it ensures the accuracy, completeness, and consistency of the dataset. In this project, various SQL queries and operations will be performed to identify and rectify errors, inconsistencies, and outliers present in the Real Estate dataset.